requirements.txt
beautifulsoup4==4.9.3
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cymem==2.0.5
en-core-web-sm==2.3.1
fastapi==0.63.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.0
idna==2.10
importlib-metadata==3.4.0
lxml==4.6.2
murmurhash==1.0.5
numpy==1.19.5
plac==1.1.3
preshed==3.0.5
priority==1.3.0
pydantic==1.7.3
python-multipart==0.0.5
readability-lxml==0.8.1
requests==2.25.1
six==1.15.0
soupsieve==2.1
spacy==2.3.5
SQLAlchemy==1.3.22
srsly==1.0.5
starlette==0.13.6
thinc==7.4.5
toml==0.10.2
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.3
wasabi==0.8.0
wsproto==1.0.0
zipp==3.4.0
